requirements.txt
# Python 3.9
# pip install -i https://pypi.tuna.tsinghua.edu.cn/simple -r requirements.txt

tqdm==4.63.1
cython==3.0.11
Pillow==9.0.1
# numpy==2.0.2
numpy==1.24.4
onnx==1.11.0
onnxruntime==1.11.0
opencv-python==4.6.0.66
insightface==0.6.2
termcolor==1.1.0
scikit-learn==1.0.2
flask==2.1.2
flask-cors==3.0.10
flasgger==0.9.5
PyMySQL==1.1.1
dbutils==1.3
# conda config --add channels https://conda.anaconda.org/qian_bi
# conda install --channel https://conda.anaconda.org/qian_bi dbutils
PyYAML==6.0.2
protobuf==3.20.1
Werkzeug==2.2.2
# Linux
# gunicorn -w 4 -b 0.0.0.0:9099 infer_server:appy
# gunicorn==23.0.0

# Windows
# waitress-serve --host=0.0.0.0 --port=9099 --threads=4 infer_server:app
waitress==3.0.0